working with int also now

+++ Sum/rmi/sum/i/src/ISum.java
@@ -1,5 +1,5 @@
 package sum.rmi.i;
 
 public interface ISum extends java.rmi.Remote {
-    String solve(long a) throws java.rmi.RemoteException;
+    String solve(int a) throws java.rmi.RemoteException;
 }
